backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pillow==10.2.0
rembg==2.0.56
numpy==1.26.3
opencv-python==4.9.0.80
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
google-generativeai==0.3.2
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
requests==2.31.0
slowapi==0.1.9
